Split latte hooks into different files
